requirements.txt
github3.py==1.3.0
boto3==1.14.35
requests==2.24.0
chalice==1.17.0